+++ src/backend/debug/urls.py
@@ -4,6 +4,7 @@ from django.urls import path
 
 from .views import (
     DebugViewHtml,
+    DebugViewNewMailboxHtml,
     DebugViewTxt,
 )
 
@@ -17,5 +18,10 @@ urlpatterns = [
         "__debug__/mail/invitation_txt",
         DebugViewTxt.as_view(),
         name="debug.mail.invitation_txt",
+    ),
+    path(
+        "__debug__/mail/new_mailbox_html",
+        DebugViewNewMailboxHtml.as_view(),
+        name="debug.mail.new_mailbox_html",
     ),
 ]